Replace all relevant dialogs with dropdown menus
