+++ app/src/main/java/com/jerboa/ui/components/comment/replynode/CommentReplyNode.kt
@@ -1,6 +1,5 @@
 package com.jerboa.ui.components.comment.replynode
 
-import android.widget.Toast
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.expandVertically
 import androidx.compose.animation.shrinkVertically
@@ -11,17 +10,13 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Bookmark
-import androidx.compose.material.icons.outlined.Block
 import androidx.compose.material.icons.outlined.BookmarkBorder
 import androidx.compose.material.icons.outlined.Comment
-import androidx.compose.material.icons.outlined.Description
-import androidx.compose.material.icons.outlined.Flag
 import androidx.compose.material.icons.outlined.Link
 import androidx.compose.material.icons.outlined.MarkChatRead
 import androidx.compose.material.icons.outlined.MarkChatUnread
 import androidx.compose.material.icons.outlined.MoreVert
 import androidx.compose.material.icons.outlined.Person
-import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.Divider
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
@@ -30,10 +25,7 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalClipboardManager
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.tooling.preview.Preview
 import com.jerboa.R
 import com.jerboa.VoteType
@@ -46,7 +38,6 @@ import com.jerboa.ui.components.comment.CommentBody
 import com.jerboa.ui.components.comment.PostAndCommunityContextHeader
 import com.jerboa.ui.components.common.ActionBarButton
 import com.jerboa.ui.components.common.CommentOrPostNodeHeader
-import com.jerboa.ui.components.common.IconAndTextDrawerItem
 import com.jerboa.ui.components.common.VoteGeneric
 import com.jerboa.ui.theme.LARGE_PADDING
 import com.jerboa.ui.theme.SMALL_PADDING
@@ -121,26 +112,15 @@ fun CommentReplyNodeInboxFooterLine(
     var showMoreOptions by remember { mutableStateOf(false) }
 
     if (showMoreOptions) {
-        CommentReplyNodeOptionsDialog(
+        CommentReplyOptionsDropdown(
             commentReplyView = commentReplyView,
             onDismissRequest = { showMoreOptions = false },
-            onPersonClick = {
-                showMoreOptions = false
-                onPersonClick(commentReplyView.creator.id)
-            },
-            onViewSourceClick = {
-                showMoreOptions = false
-                onViewSourceClick()
-            },
-            onReportClick = {
-                showMoreOptions = false
-                onReportClick(commentReplyView)
-            },
-            onBlockCreatorClick = {
-                showMoreOptions = false
-                onBlockCreatorClick(commentReplyView.creator)
-            },
+            onPersonClick = onPersonClick,
+            onViewSourceClick = onViewSourceClick,
+            onReportClick = onReportClick,
+            onBlockCreatorClick = onBlockCreatorClick,
             isCreator = account.id == commentReplyView.creator.id,
+            onCommentLinkClick = onCommentLinkClick,
             viewSource = viewSource,
         )
     }
@@ -232,76 +212,6 @@ fun CommentReplyNodeInboxFooterLine(
             )
         }
     }
-}
-
-@Composable
-fun CommentReplyNodeOptionsDialog(
-    commentReplyView: CommentReplyView,
-    onDismissRequest: () -> Unit,
-    onPersonClick: () -> Unit,
-    onViewSourceClick: () -> Unit,
-    onReportClick: () -> Unit,
-    onBlockCreatorClick: () -> Unit,
-    isCreator: Boolean,
-    viewSource: Boolean,
-) {
-    val localClipboardManager = LocalClipboardManager.current
-    val ctx = LocalContext.current
-
-    AlertDialog(
-        onDismissRequest = onDismissRequest,
-        text = {
-            Column {
-                IconAndTextDrawerItem(
-                    text = stringResource(
-                        R.string.comment_reply_node_go_to,
-                        commentReplyView.creator.name,
-                    ),
-                    icon = Icons.Outlined.Person,
-                    onClick = onPersonClick,
-                )
-                IconAndTextDrawerItem(
-                    text = if (viewSource) {
-                        stringResource(R.string.comment_node_view_original)
-                    } else {
-                        stringResource(R.string.comment_reply_node_view_source)
-                    },
-                    icon = Icons.Outlined.Description,
-                    onClick = onViewSourceClick,
-                )
-                IconAndTextDrawerItem(
-                    text = stringResource(R.string.comment_reply_node_copy_permalink),
-                    icon = Icons.Outlined.Link,
-                    onClick = {
-                        val permalink = commentReplyView.comment.ap_id
-                        localClipboardManager.setText(AnnotatedString(permalink))
-                        Toast.makeText(
-                            ctx,
-                            ctx.getString(R.string.comment_reply_node_permalink_copied),
-                            Toast.LENGTH_SHORT,
-                        ).show()
-                        onDismissRequest()
-                    },
-                )
-                if (!isCreator) {
-                    IconAndTextDrawerItem(
-                        text = stringResource(R.string.comment_reply_node_report_comment),
-                        icon = Icons.Outlined.Flag,
-                        onClick = onReportClick,
-                    )
-                    IconAndTextDrawerItem(
-                        text = stringResource(
-                            R.string.comment_reply_node_block,
-                            commentReplyView.creator.name,
-                        ),
-                        icon = Icons.Outlined.Block,
-                        onClick = onBlockCreatorClick,
-                    )
-                }
-            }
-        },
-        confirmButton = {},
-    )
 }
 
 @Composable
