+++ app/src/main/java/com/jerboa/ui/components/common/Dialogs.kt
@@ -5,8 +5,6 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.outlined.BarChart
 import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.Button
 import androidx.compose.material3.Text
@@ -20,16 +18,12 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.semantics.testTagsAsResourceId
-import androidx.compose.ui.tooling.preview.Preview
 import com.jerboa.R
 import com.jerboa.api.MINIMUM_API_VERSION
-import com.jerboa.datatypes.types.CommentSortType
-import com.jerboa.datatypes.types.SortType
 import com.jerboa.model.AppSettingsViewModel
 
 val DONATION_MARKDOWN = """
@@ -43,116 +37,6 @@ val DONATION_MARKDOWN = """
     - [Support on OpenCollective](https://opencollective.com/lemmy).
 
 """.trimIndent()
-
-val isTopSort = { sort: SortType -> sort.name.startsWith("Top") }
-
-@Composable
-fun SortTopOptionsDialog(
-    onDismissRequest: () -> Unit,
-    onClickSortType: (SortType) -> Unit,
-    selectedSortType: SortType,
-    siteVersion: String,
-) {
-    val ctx = LocalContext.current
-    AlertDialog(
-        onDismissRequest = onDismissRequest,
-        text = {
-            Column {
-                SortType.getSupportedSortTypes(siteVersion).filter(isTopSort).forEach {
-                    IconAndTextDrawerItem(
-                        text = ctx.getString(it.longForm),
-                        onClick = { onClickSortType(it) },
-                        highlight = (selectedSortType == it),
-                    )
-                }
-            }
-        },
-        confirmButton = {},
-    )
-}
-
-@Preview
-@Composable
-fun SortOptionsDialogPreview() {
-    SortOptionsDialog(
-        selectedSortType = SortType.Hot,
-        onDismissRequest = {},
-        onClickSortTopOptions = {},
-        onClickSortType = {},
-        siteVersion = MINIMUM_API_VERSION,
-    )
-}
-
-@OptIn(ExperimentalComposeUiApi::class)
-@Composable
-fun SortOptionsDialog(
-    onDismissRequest: () -> Unit,
-    onClickSortType: (SortType) -> Unit,
-    onClickSortTopOptions: () -> Unit,
-    selectedSortType: SortType,
-    siteVersion: String,
-) {
-    AlertDialog(
-        modifier = Modifier.semantics { testTagsAsResourceId = true },
-        onDismissRequest = onDismissRequest,
-        text = {
-            Column {
-                SortType.getSupportedSortTypes(siteVersion).filter { !isTopSort(it) }.forEach {
-                    IconAndTextDrawerItem(
-                        text = stringResource(it.longForm),
-                        icon = it.icon,
-                        onClick = { onClickSortType(it) },
-                        highlight = (selectedSortType == it),
-                    )
-                }
-                IconAndTextDrawerItem(
-                    text = stringResource(R.string.dialogs_top),
-                    icon = Icons.Outlined.BarChart,
-                    onClick = onClickSortTopOptions,
-                    more = true,
-                    highlight = (isTopSort(selectedSortType)),
-                )
-            }
-        },
-        confirmButton = {},
-    )
-}
-
-@Preview
-@Composable
-fun CommentSortOptionsDialogPreview() {
-    CommentSortOptionsDialog(
-        selectedSortType = CommentSortType.Hot,
-        onDismissRequest = {},
-        onClickSortType = {},
-        siteVersion = MINIMUM_API_VERSION,
-    )
-}
-
-@Composable
-fun CommentSortOptionsDialog(
-    onDismissRequest: () -> Unit,
-    onClickSortType: (CommentSortType) -> Unit,
-    selectedSortType: CommentSortType,
-    siteVersion: String,
-) {
-    AlertDialog(
-        onDismissRequest = onDismissRequest,
-        text = {
-            Column {
-                CommentSortType.getSupportedSortTypes(siteVersion).forEach {
-                    IconAndTextDrawerItem(
-                        text = stringResource(it.text),
-                        icon = it.icon,
-                        onClick = { onClickSortType(it) },
-                        highlight = (selectedSortType == it),
-                    )
-                }
-            }
-        },
-        confirmButton = {},
-    )
-}
 
 @OptIn(ExperimentalComposeUiApi::class)
 @Composable
