+++ app/src/main/java/com/jerboa/ui/components/common/LinkDropDownMenu.kt
@@ -1,10 +1,6 @@
 package com.jerboa.ui.components.common
 
-import android.content.ActivityNotFoundException
-import android.util.Log
 import android.widget.Toast
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.outlined.Download
@@ -14,10 +10,8 @@ import androidx.compose.material.icons.outlined.OpenInFull
 import androidx.compose.material.icons.outlined.Share
 import androidx.compose.material3.Divider
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalClipboardManager
 import androidx.compose.ui.platform.LocalContext
@@ -27,8 +21,6 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextDecoration
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.window.Popup
-import androidx.compose.ui.window.PopupProperties
 import com.jerboa.JerboaAppState
 import com.jerboa.PostType
 import com.jerboa.R
@@ -38,7 +30,6 @@ import com.jerboa.feat.storeMedia
 import com.jerboa.isMedia
 import com.jerboa.rememberJerboaAppState
 import com.jerboa.ui.theme.LARGE_PADDING
-import com.jerboa.ui.theme.Shapes
 
 @Composable
 fun LinkDropDownMenu(
@@ -54,164 +45,125 @@ fun LinkDropDownMenu(
     if (link != null) {
         val mediaType = PostType.fromURL(link)
 
-        Popup(
-            alignment = Alignment.Center,
+        CenteredPopupMenu(
+            expanded = true,
             onDismissRequest = onDismissRequest,
-            properties = PopupProperties(focusable = true),
+            tonalElevation = 6.dp,
         ) {
-            Surface(
-                shape = Shapes.extraSmall,
-                color = MaterialTheme.colorScheme.surface,
-                tonalElevation = 6.dp,
-                shadowElevation = 6.dp,
-            ) {
-                Column(
-                    modifier = Modifier
-                        .fillMaxWidth(0.86f)
-                        .padding(vertical = LARGE_PADDING),
-                ) {
-                    Text(
-                        text = link,
-                        textDecoration = TextDecoration.Underline,
-                        fontWeight = FontWeight.Medium,
-                        modifier = Modifier.padding(LARGE_PADDING),
-                        color = MaterialTheme.colorScheme.tertiary,
-                    )
+            Text(
+                text = link,
+                textDecoration = TextDecoration.Underline,
+                fontWeight = FontWeight.Medium,
+                modifier = Modifier.padding(LARGE_PADDING),
+                color = MaterialTheme.colorScheme.tertiary,
+            )
 
-                    MenuItem(
-                        text = stringResource(R.string.open_link),
-                        icon = Icons.Outlined.OpenInFull,
-                        onClick = {
-                            if (mediaType == PostType.Image) {
-                                appState.openImageViewer(link)
-                            } else {
-                                appState.openLink(link, useCustomTabs, usePrivateTabs)
-                            }
-                            onDismissRequest()
-                        },
-                        textStyle = MaterialTheme.typography.bodyLarge,
-                        textModifier = Modifier.padding(start = LARGE_PADDING),
-                    )
-                    MenuItem(
-                        text = stringResource(R.string.open_link_external),
-                        icon = Icons.Outlined.OpenInBrowser,
-                        onClick = {
-                            appState.openLinkRaw(link, useCustomTabs, usePrivateTabs)
+            PopupMenuItem(
+                text = stringResource(R.string.open_link),
+                icon = Icons.Outlined.OpenInFull,
+                onClick = {
+                    onDismissRequest()
+                    if (mediaType == PostType.Image) {
+                        appState.openImageViewer(link)
+                    } else {
+                        appState.openLink(link, useCustomTabs, usePrivateTabs)
+                    }
+                },
+            )
+            PopupMenuItem(
+                text = stringResource(R.string.open_link_external),
+                icon = Icons.Outlined.OpenInBrowser,
+                onClick = {
+                    onDismissRequest()
+                    appState.openLinkRaw(link, useCustomTabs, usePrivateTabs)
+                },
+            )
 
-                            try {
-                            } catch (e: ActivityNotFoundException) {
-                                Log.d("jerboa", "failed open activity", e)
-                                Toast.makeText(ctx, ctx.getText(R.string.no_activity_found), Toast.LENGTH_SHORT).show()
-                            }
+            Divider()
 
-                            onDismissRequest()
-                        },
-                        textStyle = MaterialTheme.typography.bodyLarge,
-                        textModifier = Modifier.padding(start = LARGE_PADDING),
-                    )
+            PopupMenuItem(
+                text = stringResource(R.string.post_listing_copy_link),
+                icon = Icons.Outlined.Link,
+                onClick = {
+                    onDismissRequest()
+                    localClipboardManager.setText(AnnotatedString(link))
+                    Toast.makeText(
+                        ctx,
+                        ctx.getString(R.string.post_listing_link_copied),
+                        Toast.LENGTH_SHORT,
+                    ).show()
+                },
+            )
+
+            PopupMenuItem(
+                text = stringResource(R.string.post_listing_share_link),
+                icon = Icons.Outlined.Share,
+                onClick = {
+                    onDismissRequest()
+                    shareLink(link, ctx)
+                },
+            )
+
+            when (mediaType) {
+                PostType.Image -> {
                     Divider()
-
-                    MenuItem(
-                        text = stringResource(R.string.post_listing_copy_link),
-                        icon = Icons.Outlined.Link,
-                        onClick = {
-                            localClipboardManager.setText(AnnotatedString(link))
-                            Toast.makeText(
-                                ctx,
-                                ctx.getString(R.string.post_listing_link_copied),
-                                Toast.LENGTH_SHORT,
-                            ).show()
-                            onDismissRequest()
-                        },
-                        textStyle = MaterialTheme.typography.bodyLarge,
-                        textModifier = Modifier.padding(start = LARGE_PADDING),
-                    )
-
-                    MenuItem(
-                        text = stringResource(R.string.post_listing_share_link),
+                    PopupMenuItem(
+                        text = stringResource(R.string.share_image),
                         icon = Icons.Outlined.Share,
                         onClick = {
-                            shareLink(link, ctx)
                             onDismissRequest()
+                            shareMedia(appState.coroutineScope, ctx, link, mediaType)
                         },
-                        textStyle = MaterialTheme.typography.bodyLarge,
-                        textModifier = Modifier.padding(start = LARGE_PADDING),
                     )
+                    PopupMenuItem(
+                        text = stringResource(R.string.save_image),
+                        icon = Icons.Outlined.Download,
+                        onClick = {
+                            onDismissRequest()
+                            storeMedia(appState.coroutineScope, ctx, link, mediaType)
+                        },
+                    )
+                }
 
-                    when (mediaType) {
-                        PostType.Image -> {
-                            Divider()
-                            MenuItem(
-                                text = stringResource(R.string.share_image),
-                                icon = Icons.Outlined.Share,
-                                onClick = {
-                                    shareMedia(appState.coroutineScope, ctx, link, mediaType)
-                                    onDismissRequest()
-                                },
-                                textStyle = MaterialTheme.typography.bodyLarge,
-                                textModifier = Modifier.padding(start = LARGE_PADDING),
-                            )
-                            MenuItem(
-                                text = stringResource(R.string.save_image),
-                                icon = Icons.Outlined.Download,
-                                onClick = {
-                                    storeMedia(appState.coroutineScope, ctx, link, mediaType)
-                                    onDismissRequest()
-                                },
-                                textStyle = MaterialTheme.typography.bodyLarge,
-                                textModifier = Modifier.padding(start = LARGE_PADDING),
-                            )
-                        }
+                PostType.Video -> {
+                    Divider()
+                    PopupMenuItem(
+                        text = stringResource(R.string.share_video),
+                        icon = Icons.Outlined.Share,
+                        onClick = {
+                            onDismissRequest()
+                            shareMedia(appState.coroutineScope, ctx, link, mediaType)
+                        },
+                    )
+                    PopupMenuItem(
+                        text = stringResource(R.string.save_video),
+                        icon = Icons.Outlined.Download,
+                        onClick = {
+                            onDismissRequest()
+                            storeMedia(appState.coroutineScope, ctx, link, mediaType)
+                        },
+                    )
+                }
 
-                        PostType.Video -> {
-                            Divider()
-                            MenuItem(
-                                text = stringResource(R.string.share_video),
-                                icon = Icons.Outlined.Share,
-                                onClick = {
-                                    shareMedia(appState.coroutineScope, ctx, link, mediaType)
-                                    onDismissRequest()
-                                },
-                                textStyle = MaterialTheme.typography.bodyLarge,
-                                textModifier = Modifier.padding(start = LARGE_PADDING),
-                            )
-                            MenuItem(
-                                text = stringResource(R.string.save_video),
-                                icon = Icons.Outlined.Download,
-                                onClick = {
-                                    storeMedia(appState.coroutineScope, ctx, link, mediaType)
-                                    onDismissRequest()
-                                },
-                                textStyle = MaterialTheme.typography.bodyLarge,
-                                textModifier = Modifier.padding(start = LARGE_PADDING),
-                            )
-                        }
-
-                        PostType.Link -> {
-                            if (isMedia(link)) {
-                                Divider()
-                                MenuItem(
-                                    text = stringResource(R.string.share),
-                                    icon = Icons.Outlined.Share,
-                                    onClick = {
-                                        shareMedia(appState.coroutineScope, ctx, link, PostType.Link)
-                                        onDismissRequest()
-                                    },
-                                    textStyle = MaterialTheme.typography.bodyLarge,
-                                    textModifier = Modifier.padding(start = LARGE_PADDING),
-                                )
-                                MenuItem(
-                                    text = stringResource(R.string.save),
-                                    icon = Icons.Outlined.Download,
-                                    onClick = {
-                                        storeMedia(appState.coroutineScope, ctx, link, PostType.Link)
-                                        onDismissRequest()
-                                    },
-                                    textStyle = MaterialTheme.typography.bodyLarge,
-                                    textModifier = Modifier.padding(start = LARGE_PADDING),
-                                )
-                            }
-                        }
+                PostType.Link -> {
+                    if (isMedia(link)) {
+                        Divider()
+                        PopupMenuItem(
+                            text = stringResource(R.string.share_media),
+                            icon = Icons.Outlined.Share,
+                            onClick = {
+                                onDismissRequest()
+                                shareMedia(appState.coroutineScope, ctx, link, PostType.Link)
+                            },
+                        )
+                        PopupMenuItem(
+                            text = stringResource(R.string.save),
+                            icon = Icons.Outlined.Download,
+                            onClick = {
+                                onDismissRequest()
+                                storeMedia(appState.coroutineScope, ctx, link, PostType.Link)
+                            },
+                        )
                     }
                 }
             }
