+++ app/src/main/res/values/strings.xml
@@ -411,4 +411,6 @@
     <string name="open_link">Open link</string>
     <string name="open_link_external">Open link in external</string>
     <string name="no_activity_found">No activity (app) found that can open this link</string>
+    <string name="copy">Copy</string>
+    <string name="share_media">Share media</string>
 </resources>
